requirements.txt
backcall==0.1.0
colorama==0.4.1
cycler==0.10.0
decorator==4.3.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
kiwisolver==1.0.1
matplotlib==3.0.2
mistune==0.8.4
numpy==1.15.4
olefile==0.46
opencv-python==3.4.4.19
parso==0.3.1
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
Pygments==2.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
scipy==1.1.0
six==1.11.0
traitlets==4.3.2
utils==0.9.0
wcwidth==0.1.7